feat: add jsPluginAdapter
