+++ packages/core/src/config/types.ts
@@ -2,8 +2,8 @@ import type cors from '@koa/cors';
 import type { OutgoingHttpHeaders } from 'http';
 import type { Logger } from '../utils/index.js';
 import type { ProxiesOptions } from '../server/middlewares/proxy.js';
-import type { JsPlugin } from '../plugin/index.js';
-import type { RustPlugin } from '../plugin/rustPluginResolver.js';
+import type { JsPlugin } from '../plugin/type.js';
+import type { RustPlugin } from '../plugin/rust/index.js';
 import type { Config } from '../../binding/index.js';
 import { DevServer } from '../index.js';
