+++ packages/core/src/plugin/js/index.ts
@@ -1,0 +1,1 @@
+export * from './jsPluginAdapter.js';
