+++ packages/core/src/plugin/rust/index.ts
@@ -1,0 +1,1 @@
+export * from './rustPluginResolver.js';
