+++ packages/core/src/plugin/rust/rustPluginResolver.ts
